+++ src/main/java/com/bookbla/americano/domain/library/controller/MemberLibraryController.java
@@ -1,10 +1,10 @@
-package com.bookbla.americano.domain.member.controller;
+package com.bookbla.americano.domain.library.controller;
 
 import com.bookbla.americano.base.resolver.LoginUser;
 import com.bookbla.americano.base.resolver.User;
-import com.bookbla.americano.domain.member.controller.dto.response.MemberLibraryProfileReadResponse;
-import com.bookbla.americano.domain.member.controller.dto.response.MemberTargetLibraryProfileReadResponse;
-import com.bookbla.americano.domain.member.service.MemberLibraryService;
+import com.bookbla.americano.domain.library.service.MemberLibraryService;
+import com.bookbla.americano.domain.library.controller.dto.MemberLibraryProfileReadResponse;
+import com.bookbla.americano.domain.library.controller.dto.MemberTargetLibraryProfileReadResponse;
 import io.swagger.v3.oas.annotations.Parameter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
